==> notebooks/Step1_Data collection.ipynb <==
import pandas as pd

# Load the dataset
df = pd.read_csv("UserFeedbackData.csv")

# Quick check of the first 5 rows
df.head()

# Filter and check shape
spotify_df = df[df["app_name"] == "Spotify"].copy()
print(f"Dataset Shape: {spotify_df.shape}")


reviews_per_version = (
    spotify_df
    .groupby("RC_ver")
    .size()
    .sort_values(ascending=False)
)
print("--- Reviews per Version ---")
reviews_per_version

avg_rating_per_version = (
    spotify_df
    .groupby("RC_ver")["score"]
   .agg(['mean', 'size'])
    .sort_values(by="mean", ascending=False)
)

print("--- Average Rating per Version ---")
avg_rating_per_version

rating_distribution = spotify_df["score"].value_counts().sort_index()
print("--- Rating Distribution ---")
rating_distribution

# Set the max column width to None (unlimited)
pd.set_option('display.max_colwidth', None)

top_liked_reviews = (
    spotify_df
    .sort_values("TU_count", ascending=False)
    .loc[:, ["RC_ver", "score", "TU_count", "content"]]
    .head(10)
)

print("--- Top Liked Reviews ---")
top_liked_reviews

low_star_reviews = spotify_df[spotify_df["score"] <= 2]

low_star_by_version = (
    low_star_reviews
    .groupby("RC_ver")
    .size()
    .sort_values(ascending=False)
)

print("--- Low Star Reviews by Version ---")
low_star_by_version
